pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "swordfish_extimageapi"
version = "0.0.1"
description = "A collection of custom nodes for ComfyUI"
authors = [
  {name = "Enzo Casagrande", email = "enzo.casagrande@silverside.ai"}
]
readme = "README.md"
license = {text = "Not open source"}
requires-python = ">=3.10"
classifiers = []
dependencies = [
    "google-genai",
    "Pillow",
]



[project.optional-dependencies]
dev = [
    "bump-my-version",
    "coverage",  # testing
    "mypy",  # linting
    "pre-commit", # runs linting on commit
    "pytest",  # testing
    "ruff",  # linting
]

[project.urls]
Repository = "https://github.com/enzocasagrande-silverside/swordfish_extimageapi"
BugTracker = "https://github.com/enzocasagrande-silverside/swordfish_extimageapi/issues"
Documentation = "https://github.com/enzocasagrande-silverside/swordfish_extimageapi/wiki"


[tool.comfy]
PublisherId = "enzocasagrande-silverside"
DisplayName = "swordfish-extimageapi"
Icon = ""
Tags = []
Repository = "https://github.com/enzocasagrande-silverside/swordfish_extimageapi"

includes = []

[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
]

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

[tool.ruff]
# extend-exclude = ["static", "ci/templates"]
line-length = 140
src = ["src", "tests"]
target-version = "py39"

# Add rules to ban exec/eval
[tool.ruff.lint]
select = [
    "S102",  # exec-builtin
    "S307",  # eval-used
    "W293",
    "F",  # The "F" series in Ruff stands for "Pyflakes" rules, which catch various Python syntax errors and undefined names.
    # See all rules here: https://docs.astral.sh/ruff/rules/#pyflakes-f
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
